更新服务端Java源码(新增功能与服务层实现)

+++ gms-server/src/main/java/org/gms/client/inventory/InventoryProof.java
@@ -27,7 +27,7 @@ import org.gms.client.Character;
 public class InventoryProof extends Inventory {
 
     public InventoryProof(Character mc) {
-        super(mc, InventoryType.CANHOLD, (byte) 0);
+        super(mc, InventoryType.CANHOLD, (short) 0);
     }
 
     public void cloneContents(Inventory inv) {
